linted and restructured
