--- mobile/components/index.ts
@@ -1,3 +1,0 @@
-export { default as PostCardPlaceholder } from './controls/PostCardPlaceholder';
-export { default as PostCard } from './controls/PostCard';
-export { default as HomeScreenFeed } from './controls/HomeScreenFeed';
