--- mobile/components/presentational/index.ts
@@ -1,8 +1,0 @@
-export { default as PVActionButton } from './PVActionButton';
-export { default as PVFormInput } from './PVFormInput';
-export { default as PVFormToggle } from './PVFormToggle';
-export { default as PVIcon } from './PVIcon';
-export { default as PVImageButton } from './PVImageButton';
-export { default as PVTabBarButton } from './PVTabBarButton';
-export { default as PostCard } from '../controls/PostCard';
-export { default as PostCardPlaceHolder } from '../controls/PostCardPlaceholder';
