+++ mobile/navigation/index.ts
@@ -1,4 +1,4 @@
-export { default as NVAuthStackLayout } from './NVAuthStackLayout';
-export { default as NVFriendStackLayout } from './NVFriendStackLayout';
-export { default as NVMainTabLayout } from './NVMainTabLayout';
-export { default as NVRootLayout } from './NVRootLayout';
+export { default as NVAuthStackLayout } from "./NVAuthStackLayout";
+export { default as NVFriendStackLayout } from "./NVFriendStackLayout";
+export { default as NVMainTabLayout } from "./NVMainTabLayout";
+export { default as NVRootLayout } from "./NVRootLayout";
