+++ mobile/tsconfig.json
@@ -4,16 +4,14 @@
     "baseUrl": "./",
     "strict": true,
     "paths": {
-      "@components": ["components/index"],
       "@components/*": ["components/*"],
       "@containers/*": ["containers/*"],
       "@constants": ["constants/index"],
       "@constants/*": ["constants/*"],
-      "@context": ["context/index"],
+      "@context": ["context"],
       "@context/*": ["context/*"],
       "@hooks/*": ["hooks/*"],
       "@utils/*": ["utils/*"],
-      "@utils/": ["utils/index"],
       "@navigation/*": ["navigation/*"],
       "@services/*": ["services/*"],
       "@env": ["node_modules/react-native-dotenv"]
