+++ mobile/utils/index.ts
@@ -1,2 +1,3 @@
-export * from './Logger';
-export * from './helpers';
+export * from "./Logger";
+export * from "./Sentry";
+export * from "./helpers";
